requirements.txt
pykeepass
PyYAML
requests